[Refactor] file hierarachy
