+++ github-issue/src/App.js
@@ -1,6 +1,6 @@
 import './App.css';
-import Header from './components/Header';
-import Main from './components/Main';
+import Header from './components/Layout/Header';
+import Main from './components/Layout/Main';
 
 function App() {
   return (
